backend/requirements.txt
# Jarvis Assistant - Backend Requirements
# Python 3.10+
# Install with: pip install -r requirements.txt

# ============================================
# CORE - Web Framework & Utilities (REQUIRED)
# ============================================
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
httpx>=0.25.2
aiofiles>=23.2.1

# ============================================
# DATABASE (REQUIRED)
# ============================================
pymongo>=4.6.0

# ============================================
# AI / LLM (REQUIRED)
# ============================================
openai>=1.3.0

# ============================================
# PDF Processing (REQUIRED)
# ============================================
PyMuPDF>=1.23.8

# ============================================
# Scheduling (REQUIRED)
# ============================================
APScheduler>=3.10.4

# ============================================
# Testing (REQUIRED)
# ============================================
pytest>=7.4.3
pytest-asyncio>=0.21.1

# ============================================
# ONLINE TTS (RECOMMENDED)
# ============================================
gtts>=2.5.0
pygame>=2.5.0

# ============================================
# OFFLINE TTS - Lightweight (RECOMMENDED)
# ============================================
pyttsx3>=2.90

# ============================================
# OFFLINE STT (OPTIONAL - 300MB download)
# ============================================
# faster-whisper>=0.10.0

# ============================================
# OFFLINE TTS - High Quality (OPTIONAL - 1GB)
# ============================================
# TTS>=0.22.0

# ============================================
# WAKE WORD DETECTION (OPTIONAL - needs key)
# ============================================
# pvporcupine>=3.0.0
pyaudio>=0.2.14  # Microphone recording (REQUIRED FOR STT)

# ============================================
# EMAIL INTEGRATION (OPTIONAL)
# ============================================
# google-auth-oauthlib>=1.1.0
# google-auth-httplib2>=0.1.1
# google-api-python-client>=2.108.0

# ============================================
# ALTERNATIVE STT (OPTIONAL)
# ============================================
# vosk>=0.3.45
# openai-whisper>=20231117

# ============================================
# AUDIO PROCESSING (OPTIONAL)
# ============================================
# sounddevice>=0.4.6
# soundfile>=0.12.1
# pydub>=0.25.1
